Use docstrings when documenting functions+classes
There's a multi-line version which uses 3x double-quotes too; the key
thing is to put the comment as the first line in the function/class.
It's also important to stick with the 80-char limit, especially for the
first line.  PEP-8 is the usual style guide for Python, you don't really
need to follow it religiously but do stick with it when you can.